pyproject.toml
[project]
name = "proto-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.5",
    "motor>=3.6.0",
    "orjson>=3.10.12",
    "pycryptodome>=3.21.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
]
